% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{as_hex}
\alias{as_hex}
\title{as_hex
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
as_hex(inGeom)

\S4method{as_hex}{Column}(inGeom)
}
\description{
as_hex
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
