% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as_json}
\alias{as_json}
\title{as_json
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{as_json}{Column}(inGeom)
}
\description{
as_json
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
