% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enableMosaic.R
\name{enableMosaic}
\alias{enableMosaic}
\title{enableMosaic}
\usage{
enableMosaic(geometryAPI = "ESRI", indexSystem = "H3")
}
\arguments{
\item{geometryAPI}{character, default="ESRI"}

\item{indexSystem}{character, default="H3"}
}
\value{
None
}
\description{
enableMosaic activates the context dependent Databricks Mosaic functions, giving control over the geometry API and index system used.
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
\examples{
\dontrun{
enableMosaic()
enableMosaic("ESRI", "H3")
enableMosaic("ESRI", "BNG") # Not yet supported}
}
