% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{index_geometry}
\alias{index_geometry}
\title{index_geometry
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{index_geometry}{Column}(indexID)
}
\description{
index_geometry
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
