% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{mosaic_explode}
\alias{mosaic_explode}
\alias{mosaic_explode,Column,Column-method}
\alias{mosaic_explode,Column,numeric-method}
\title{mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
mosaic_explode(geom, resolution)

mosaic_explode(geom, resolution)

\S4method{mosaic_explode}{Column,Column}(geom, resolution)

\S4method{mosaic_explode}{Column,numeric}(geom, resolution)
}
\arguments{
\item{geom}{Column}

\item{resolution}{numeric}
}
\description{
mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
