% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{point_index_geom}
\alias{point_index_geom}
\title{point_index_geom
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{point_index_geom}{Column,Column}(point, resolution)

\S4method{point_index_geom}{Column,numeric}(point, resolution)
}
\description{
point_index_geom
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

point_index_geom
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
