% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{polyfill}
\alias{polyfill}
\title{polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{polyfill}{Column,Column}(geom, resolution)

\S4method{polyfill}{Column,numeric}(geom, resolution)
}
\description{
polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
