% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_asbinary}
\alias{st_asbinary}
\title{st_asbinary
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_asbinary}{Column}(geom)
}
\description{
st_asbinary
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
