% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_asgeojson}
\alias{st_asgeojson}
\title{st_asgeojson
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_asgeojson(geom)

\S4method{st_asgeojson}{Column}(geom)
}
\description{
st_asgeojson
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
