% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_aswkb}
\alias{st_aswkb}
\title{st_aswkb
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_aswkb}{Column}(geom)
}
\description{
st_aswkb
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
