% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_centroid3D}
\alias{st_centroid3D}
\title{st_centroid3D
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_centroid3D(geom)

\S4method{st_centroid3D}{Column}(geom)
}
\description{
st_centroid3D
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
