% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_dump}
\alias{st_dump}
\title{st_dump
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_dump}{Column}(geom)
}
\description{
st_dump
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
