% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_geomfromwkb}
\alias{st_geomfromwkb}
\title{st_geomfromwkb
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_geomfromwkb}{Column}(inGeom)
}
\description{
st_geomfromwkb
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
