% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_geomfromwkt}
\alias{st_geomfromwkt}
\title{st_geomfromwkt
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_geomfromwkt(inGeom)

\S4method{st_geomfromwkt}{Column}(inGeom)
}
\description{
st_geomfromwkt
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
