% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_intersection_aggregate}
\alias{st_intersection_aggregate}
\title{st_intersection_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_intersection_aggregate}{Column,Column}(leftIndex, rightIndex)
}
\description{
st_intersection_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
