% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_intersects_aggregate}
\alias{st_intersects_aggregate}
\title{st_intersects_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_intersects_aggregate(leftIndex, rightIndex)

\S4method{st_intersects_aggregate}{Column,Column}(leftIndex, rightIndex)
}
\description{
st_intersects_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
