% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_isvalid}
\alias{st_isvalid}
\title{st_isvalid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_isvalid}{Column}(geom)
}
\description{
st_isvalid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
