% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_makepolygon}
\alias{st_makepolygon}
\title{st_makepolygon
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_makepolygon}{Column,ANY}(boundaryRing)

\S4method{st_makepolygon}{Column,Column}(boundaryRing, holeRingArray)
}
\description{
st_makepolygon
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

st_makepolygon
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
