% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_perimeter}
\alias{st_perimeter}
\title{st_perimeter
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_perimeter}{Column}(geom)
}
\description{
st_perimeter
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
