% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_point}
\alias{st_point}
\title{st_point
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_point}{Column,Column}(xVal, yVal)
}
\description{
st_point
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
