% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_rotate}
\alias{st_rotate}
\alias{st_rotate,Column,Column-method}
\title{st_rotate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_rotate(geom1, td)

\S4method{st_rotate}{Column,Column}(geom1, td)
}
\arguments{
\item{geom1}{Column}

\item{td}{Column}
}
\description{
st_rotate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
