% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_scale}
\alias{st_scale}
\title{st_scale
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_scale(geom1, xd, yd)

\S4method{st_scale}{Column,Column,Column}(geom1, xd, yd)
}
\description{
st_scale
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
