% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_setsrid}
\alias{st_setsrid}
\alias{st_setsrid,Column,Column-method}
\title{st_setsrid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_setsrid(geom, srid)

\S4method{st_setsrid}{Column,Column}(geom, srid)
}
\arguments{
\item{geom}{Column}

\item{srid}{Column}
}
\description{
st_setsrid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
