% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_transform}
\alias{st_transform}
\title{st_transform
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_transform}{Column,Column}(geom, srid)
}
\description{
st_transform
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
