% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_translate}
\alias{st_translate}
\title{st_translate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_translate}{Column,Column,Column}(geom1, xd, yd)
}
\description{
st_translate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
