% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_xmax}
\alias{st_xmax}
\title{st_xmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_xmax}{Column}(geom)
}
\description{
st_xmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
