% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_xmin}
\alias{st_xmin}
\title{st_xmin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_xmin(geom)

\S4method{st_xmin}{Column}(geom)
}
\description{
st_xmin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
