% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_ymax}
\alias{st_ymax}
\title{st_ymax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_ymax}{Column}(geom)
}
\description{
st_ymax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
