% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_ymin}
\alias{st_ymin}
\title{st_ymin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_ymin}{Column}(geom)
}
\description{
st_ymin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
