% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_zmax}
\alias{st_zmax}
\alias{st_zmax,Column-method}
\title{st_zmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_zmax(geom)

\S4method{st_zmax}{Column}(geom)
}
\arguments{
\item{geom}{Column}
}
\description{
st_zmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
