% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{try_sql}
\alias{try_sql}
\title{try_sql
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{try_sql}{Column}(inCol)
}
\description{
try_sql
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
