% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_to}
\alias{convert_to}
\title{convert_to
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{convert_to}{Column,character}(inGeom, outDataType)
}
\description{
convert_to
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
