% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flatten_polygons}
\alias{flatten_polygons}
\title{flatten_polygons
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{flatten_polygons}{Column}(geom)
}
\description{
flatten_polygons
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
