% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mosaic_explode}
\alias{mosaic_explode}
\title{mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{mosaic_explode}{Column,Column}(geom, resolution)

\S4method{mosaic_explode}{Column,numeric}(geom, resolution)
}
\description{
mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

mosaic_explode
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
