% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{point_index_lonlat}
\alias{point_index_lonlat}
\title{point_index_lonlat
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{point_index_lonlat}{Column,Column,Column}(lon, lat, resolution)

\S4method{point_index_lonlat}{Column,Column,numeric}(lon, lat, resolution)
}
\description{
point_index_lonlat
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

point_index_lonlat
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
