% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{polyfill}
\alias{polyfill}
\alias{polyfill,Column,Column-method}
\alias{polyfill,Column,numeric-method}
\title{polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
polyfill(geom, resolution)

polyfill(geom, resolution)

\S4method{polyfill}{Column,Column}(geom, resolution)

\S4method{polyfill}{Column,numeric}(geom, resolution)
}
\arguments{
\item{geom}{Column}

\item{resolution}{numeric}
}
\description{
polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation

polyfill
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
