% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_area}
\alias{st_area}
\title{st_area
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_area(geom)

\S4method{st_area}{Column}(geom)
}
\description{
st_area
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
