% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_centroid2D}
\alias{st_centroid2D}
\title{st_centroid2D
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_centroid2D}{Column}(geom)
}
\description{
st_centroid2D
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
