% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_convexhull}
\alias{st_convexhull}
\title{st_convexhull
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_convexhull}{Column}(geom)
}
\description{
st_convexhull
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
