% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_distance}
\alias{st_distance}
\title{st_distance
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_distance(geom1, geom2)

\S4method{st_distance}{Column,Column}(geom1, geom2)
}
\description{
st_distance
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
