% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_intersection_aggregate}
\alias{st_intersection_aggregate}
\alias{st_intersection_aggregate,Column,Column-method}
\title{st_intersection_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_intersection_aggregate(leftIndex, rightIndex)

\S4method{st_intersection_aggregate}{Column,Column}(leftIndex, rightIndex)
}
\arguments{
\item{leftIndex}{Column}

\item{rightIndex}{Column}
}
\description{
st_intersection_aggregate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
