% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_makeline}
\alias{st_makeline}
\title{st_makeline
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_makeline}{Column}(points)
}
\description{
st_makeline
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
