% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_numpoints}
\alias{st_numpoints}
\title{st_numpoints
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_numpoints(geom)

\S4method{st_numpoints}{Column}(geom)
}
\description{
st_numpoints
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
