% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_rotate}
\alias{st_rotate}
\title{st_rotate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_rotate}{Column,Column}(geom1, td)
}
\description{
st_rotate
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
