% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_srid}
\alias{st_srid}
\title{st_srid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_srid}{Column}(geom)
}
\description{
st_srid
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
