% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_xmax}
\alias{st_xmax}
\alias{st_xmax,Column-method}
\title{st_xmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_xmax(geom)

\S4method{st_xmax}{Column}(geom)
}
\arguments{
\item{geom}{Column}
}
\description{
st_xmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
