% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\name{st_ymax}
\alias{st_ymax}
\alias{st_ymax,Column-method}
\title{st_ymax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
st_ymax(geom)

\S4method{st_ymax}{Column}(geom)
}
\arguments{
\item{geom}{Column}
}
\description{
st_ymax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
