% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_zmax}
\alias{st_zmax}
\title{st_zmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_zmax}{Column}(geom)
}
\description{
st_zmax
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
