% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_zmin}
\alias{st_zmin}
\title{st_zmin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation}
\usage{
\S4method{st_zmin}{Column}(geom)
}
\description{
st_zmin
See \url{https://databrickslabs.github.io/mosaic/} for full documentation
}
